more refactoring and handles virtual study saving in modal
