+++ src/components/CohortBuilder/VirtualStudyListProvider.js
@@ -1,29 +1,9 @@
 import React from 'react';
 import Component from 'react-component-component';
 import { compose } from 'recompose';
-import gql from 'graphql-tag';
-import { print } from 'graphql/language/printer';
-import urlJoin from 'url-join';
 
 import { withApi } from 'services/api';
-import { personaApiRoot } from 'common/injectGlobals';
-
-export const getSavedVirtualStudyNames = async api =>
-  api({
-    url: urlJoin(personaApiRoot, 'graphql', 'PERSONA_SAVED_VIRTUAL_STUDIES'),
-    body: {
-      query: print(gql`
-        {
-          self {
-            virtualStudies {
-              id
-              name
-            }
-          }
-        }
-      `),
-    },
-  });
+import { getSavedVirtualStudyNames } from './utils';
 
 export default compose(withApi)(({ api, onUpdate = () => {}, children }) => {
   const initialState = {
